requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
pyyaml>=6.0.1
lxml>=5.0.0
